feat: start game button works
